Add more ducks
