+++ static/js/CustomMessageHandler.js
@@ -3,6 +3,8 @@ import * as Action from './actions';
 import * as sourceActions from './ducks/source';
 import * as fetchGroupActions from './ducks/fetchGroup';
 import * as groupsActions from './ducks/groups';
+import * as userProfileActions from './ducks/userProfile';
+import * as fetchSources from './ducks/fetchSources';
 
 
 const CustomMessageHandler = (dispatch, getState) => (
@@ -36,7 +38,7 @@ const CustomMessageHandler = (dispatch, getState) => (
         break;
       }
       case Action.FETCH_USER_PROFILE: {
-        dispatch(Action.fetchUserProfile());
+        dispatch(userProfileActions.fetchUserProfile());
         break;
       }
       default:
