+++ static/js/ducks/userProfile.js
@@ -1,0 +1,18 @@
+import * as API from '../API';
+
+
+export const FETCH_USER_PROFILE = 'skyportal/FETCH_USER_PROFILE';
+export const FETCH_USER_PROFILE_OK = 'skyportal/FETCH_USER_PROFILE_OK';
+
+export function fetchUserProfile() {
+  return API.GET('/api/internal/profile', FETCH_USER_PROFILE);
+}
+
+export default function reducer(state={ username: '', roles: [], acls: [], tokens: [] }, action) {
+  switch (action.type) {
+    case FETCH_USER_PROFILE_OK:
+      return action.data;
+    default:
+      return state;
+  }
+}
